Modelo/requirements.txt
affine==2.4.0
annotated-types==0.6.0
anyio==3.7.1
appdirs==1.4.4
asttokens==2.4.1
attrs==23.1.0
Automat==0.8.0
azure-core==1.29.5
azure-identity==1.15.0
backcall==0.2.0
blinker==1.4
cattrs==23.1.2
certifi==2019.11.28
chardet==3.0.4
click==8.1.7
click-plugins==1.1.1
cligj==0.7.2
cloud-init==23.3.1
colorama==0.4.3
comm==0.2.0
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
contourpy==1.1.1
cryptography==2.8
cycler==0.12.1
dbus-python==1.2.16
debugpy==1.8.0
decorator==5.1.1
distro==1.4.0
distro-info==0.23+ubuntu1.1
earthpy==0.9.4
entrypoints==0.3
exceptiongroup==1.1.3
executing==2.0.1
fastapi==0.104.1
fiona==1.9.5
Flask==1.1.1
fonttools==4.44.0
geojson-rewind==1.1.0
geomet==1.0.0
geopandas==0.13.2
h11==0.14.0
httplib2==0.14.0
httptools==0.6.1
hyperlink==19.0.0
idna==2.8
imageio==2.33.0
importlib-metadata==6.8.0
importlib-resources==6.1.0
incremental==16.10.1
ipykernel==6.26.0
ipython==8.12.3
itsdangerous==1.1.0
jedi==0.19.1
Jinja2==2.10.1
joblib==1.3.2
jsonpatch==1.22
jsonpointer==2.0
jsonschema==3.2.0
jupyter-client==8.6.0
jupyter-core==5.5.0
keyring==18.0.1
kiwisolver==1.4.5
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lazy-loader==0.3
lightgbm==4.1.0
MarkupSafe==1.1.0
matplotlib==3.7.3
matplotlib-inline==0.1.6
more-itertools==4.2.0
msal==1.25.0
msal-extensions==1.0.0
nest-asyncio==1.5.8
netifaces==0.10.4
networkx==3.1
numpy==1.24.4
oauthlib==3.1.0
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==10.1.0
platformdirs==3.11.0
portalocker==2.8.2
prompt-toolkit==3.0.41
psutil==5.9.6
pure-eval==0.2.2
pyasn1==0.4.2
pyasn1-modules==0.2.1
pydantic==2.4.2
pydantic-core==2.10.1
pygbif==0.6.3
pygments==2.17.1
PyGObject==3.36.0
PyHamcrest==1.9.0
pyimpute==0.3
pyinotify==0.9.6
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyodbc==5.0.1
pyOpenSSL==19.0.0
pyparsing==3.1.1
pyparted==3.11.2
pyproj==3.5.0
pyrsistent==0.15.5
pyserial==3.4
python-apt==2.0.1+ubuntu0.20.4.1
python-dateutil==2.8.2
python-debian==0.1.36+ubuntu1.1
python-dotenv==1.0.0
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML==5.3.1
pyzmq==25.1.1
rasterio==1.3.9
rasterstats==0.19.0
requests==2.22.0
requests-cache==1.1.0
requests-unixsocket==0.2.0
rioxarray==0.13.4
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
seaborn==0.13.0
SecretStorage==2.3.1
service-identity==18.1.0
shapely==2.0.2
simplejson==3.16.0
six==1.14.0
sniffio==1.3.0
snuggs==1.4.7
sos==4.5.6
ssh-import-id==5.10
stack-data==0.6.3
starlette==0.27.0
systemd-python==234
threadpoolctl==3.2.0
tifffile==2023.7.10
tornado==6.3.3
traitlets==5.13.0
Twisted==18.9.0
typing-extensions==4.8.0
tzdata==2023.3
ubuntu-advantage-tools==8001
ufw==0.36
unattended-upgrades==0.1
url-normalize==1.4.3
urllib3==1.25.8
uvicorn==0.24.0
uvloop==0.19.0
wadllib==1.3.3
WALinuxAgent==2.2.46
watchfiles==0.21.0
wcwidth==0.2.10
websockets==12.0
Werkzeug==0.16.1
xarray==2023.1.0
xgboost==2.0.2
zipp==3.17.0
zope.interface==4.7.1
